# Replace weighted semaphore with channel