--- a/x/merkledb/db.go
+++ b/x/merkledb/db.go
@@ -15,7 +15,6 @@ import (
 	"github.com/prometheus/client_golang/prometheus"
 	"go.opentelemetry.io/otel/attribute"
 	"golang.org/x/exp/maps"
-	"golang.org/x/sync/semaphore"
 
 	"github.com/ava-labs/avalanchego/database"
 	"github.com/ava-labs/avalanchego/ids"
@@ -218,7 +217,7 @@ type merkleDB struct {
 
 	// calculateNodeIDsSema controls the number of goroutines inside
 	// [calculateNodeIDsHelper] at any given time.
-	calculateNodeIDsSema *semaphore.Weighted
+	calculateNodeIDsSema semaphore
 
 	tokenSize int
 }
@@ -274,7 +273,7 @@ func newDatabase(
 		debugTracer:          getTracerIfEnabled(config.TraceLevel, DebugTrace, config.Tracer),
 		infoTracer:           getTracerIfEnabled(config.TraceLevel, InfoTrace, config.Tracer),
 		childViews:           make([]*view, 0, defaultPreallocationSize),
-		calculateNodeIDsSema: semaphore.NewWeighted(int64(rootGenConcurrency)),
+		calculateNodeIDsSema: make(semaphore, rootGenConcurrency),
 		tokenSize:            BranchFactorToTokenSize[config.BranchFactor],
 	}