--- a/x/merkledb/semaphore.go
+++ b/x/merkledb/semaphore.go
@@ -1,0 +1,27 @@
+// Copyright (C) 2019-2024, Ava Labs, Inc. All rights reserved.
+// See the file LICENSE for licensing terms.
+
+package merkledb
+
+type semaphore chan struct{}
+
+func (s semaphore) Acquire() {
+	s <- struct{}{}
+}
+
+func (s semaphore) TryAcquire() bool {
+	select {
+	case s <- struct{}{}:
+		return true
+	default:
+		return false
+	}
+}
+
+func (s semaphore) Release() {
+	select {
+	case <-s:
+	default:
+		panic("release of unacquired semaphore")
+	}
+}