--- a/x/merkledb/semaphore_test.go
+++ b/x/merkledb/semaphore_test.go
@@ -3,9 +3,7 @@
 
 package merkledb
 
-import (
-	"testing"
-)
+import "testing"
 
 func Benchmark_Semaphore_Acquire(b *testing.B) {
 	s := make(semaphore, b.N)