--- a/x/merkledb/view.go
+++ b/x/merkledb/view.go
@@ -244,9 +244,9 @@ func (v *view) calculateNodeIDs(ctx context.Context) error {
 		}
 
 		if !v.root.IsNothing() {
-			_ = v.db.calculateNodeIDsSema.Acquire(context.Background(), 1)
+			v.db.calculateNodeIDsSema.Acquire()
 			v.changes.rootID = v.calculateNodeIDsHelper(v.root.Value())
-			v.db.calculateNodeIDsSema.Release(1)
+			v.db.calculateNodeIDsSema.Release()
 		} else {
 			v.changes.rootID = ids.Empty
 		}
@@ -282,11 +282,11 @@ func (v *view) calculateNodeIDsHelper(n *node) ids.ID {
 		childEntry.hasValue = childNodeChange.after.hasValue()
 
 		// Try updating the child and its descendants in a goroutine.
-		if ok := v.db.calculateNodeIDsSema.TryAcquire(1); ok {
+		if v.db.calculateNodeIDsSema.TryAcquire() {
 			wg.Add(1)
 			go func() {
 				childEntry.id = v.calculateNodeIDsHelper(childNodeChange.after)
-				v.db.calculateNodeIDsSema.Release(1)
+				v.db.calculateNodeIDsSema.Release()
 				wg.Done()
 			}()
 		} else {