## Why this should be merged

Before:
```
Benchmark_Semaphore_Acquire-12                	79965123	        15.26 ns/op	       0 B/op	       0 allocs/op
Benchmark_Semaphore_Release-12                	77948438	        15.40 ns/op	       0 B/op	       0 allocs/op
Benchmark_Semaphore_TryAcquire_Success-12     	78838228	        15.01 ns/op	       0 B/op	       0 allocs/op
Benchmark_Semaphore_TryAcquire_Failure-12     	99081432	        11.77 ns/op	       0 B/op	       0 allocs/op
```

After:
```
Benchmark_Semaphore_Acquire-12               	56607668	        20.70  ns/op	       0 B/op	       0 allocs/op
Benchmark_Semaphore_Release-12               	55961013	        21.51  ns/op	       0 B/op	       0 allocs/op
Benchmark_Semaphore_TryAcquire_Success-12    	56662573	        21.54  ns/op	       0 B/op	       0 allocs/op
Benchmark_Semaphore_TryAcquire_Failure-12    	318408416	         3.779 ns/op	       0 B/op	       0 allocs/op
```

Because `TryAcquire_Failure` is the common case when hashing, we should optimize this path.

## How this works

Replaces the `semaphore` package with a simple wrapper around a channel.

## How this was tested

- [X] New benchmark
- [X] CI