@StephenButtolph 2024-04-01
We can wait until #2899 is merged to get a better e2e benchmark