@StephenButtolph 2024-04-01
(Running benchmarks on top of #2899, this does seem to be an improvement for larger tries when parallelism is enabled:

Before:
```
Benchmark_HashChangedNodes/1-12  		 3299204	       358.3 ns/op	     144 B/op	       3 allocs/op
Benchmark_HashChangedNodes/10-12 		   58080	     20863   ns/op	     721 B/op	      25 allocs/op
Benchmark_HashChangedNodes/100-12         	   10000	    114356   ns/op	    2872 B/op	     111 allocs/op
Benchmark_HashChangedNodes/1000-12        	    1384	    912711   ns/op	   17346 B/op	     813 allocs/op
Benchmark_HashChangedNodes/10000-12       	     152	   7474488   ns/op	   91583 B/op	    4331 allocs/op
Benchmark_HashChangedNodes/100000-12      	      14	  79664562   ns/op	 1156965 B/op	   65749 allocs/op
```

After:
```
Benchmark_HashChangedNodes/1-12  		 2710024	       400.2 ns/op	     144 B/op	       3 allocs/op
Benchmark_HashChangedNodes/10-12 		   52284	     22970   ns/op	     722 B/op	      25 allocs/op
Benchmark_HashChangedNodes/100-12         	    8569	    139937   ns/op	    2888 B/op	     111 allocs/op
Benchmark_HashChangedNodes/1000-12        	    1370	    926422   ns/op	   15252 B/op	     739 allocs/op
Benchmark_HashChangedNodes/10000-12       	     172	   7068256   ns/op	   89346 B/op	    4253 allocs/op
Benchmark_HashChangedNodes/100000-12      	      14	  75760193   ns/op	 1060367 B/op	   62297 allocs/op
```

It is a minor performance degradation for smaller tries as more of the `TryAcquire` calls `succeed` which is slower with this implementation.